+++ scripts/Battle/Statuses/DealDamageGroup.cs
@@ -7,7 +7,7 @@ public class DealDamageGroup : StatusGroup
 {
     // Wrapper
     // Start is called before the first frame update
-    public DealDamageGroup(GameObject from, GameObject target, int dmg) :
+    public DealDamageGroup(GameObject from, GameObject target, float dmg) :
         base(from, target)
     {
         name = "DealDamage";
